Delete some warning at compiling time.
git-svn-id: https://svn.csie.net/chewing/libchewing/branches/win32-utf8@646 ac1be623-90ea-0310-9410-ba68b2efa777

+++ src/zuin.c
@@ -401,7 +401,7 @@ int PinYingInput( ZuinData *pZuin, int key )
 			return ZUIN_KEY_ERROR;
 
 		DEBUG_OUT( "zuinKeySeq: %s\n", zuinKeySeq );
-		for ( i = 0; i < strlen( zuinKeySeq ); i++ ) {
+		for ( i = 0; i < (int) strlen( zuinKeySeq ); i++ ) {
 			status = DefPhoInput( pZuin, zuinKeySeq[ i ] );
 			if ( status != ZUIN_ABSORB )
 				return ZUIN_KEY_ERROR;
